refactor(pool): tests
