+++ attestor/attestor/Cargo.toml
@@ -88,5 +88,10 @@ user = { workspace = true }
 
 [dev-dependencies]
 assert_matches = { workspace = true }
+proptest = { workspace = true }
+proptest-state-machine = { workspace = true }
 rstest = { workspace = true }
 tokio-test = { workspace = true }
+
+[features]
+simulation = []
